+++ birdnet_stm32/evaluation/metrics.py
@@ -35,7 +35,7 @@ def make_chunks_for_file(
         List of per-chunk inputs ready for model.predict.
     """
     sr = int(cfg["sample_rate"])
-    cd = int(cfg["chunk_duration"])
+    cd = float(cfg["chunk_duration"])
     num_mels = int(cfg["num_mels"])
     spec_width = int(cfg["spec_width"])
